requirements.txt
dotenv
discord.py
discord.py[voice]
PyNaCl
yt-dlp
